+++ src/app/components/header/Nav.tsx
@@ -1,0 +1,31 @@
+import React from "react";
+import Link from "next/link";
+import "@/app/styles/components/Header.scss";
+
+function Nav({ isburgerNav }) {
+  return (
+    <ul className={isburgerNav ? "nav__menu__burger" : "nav__menu"}>
+      <li>
+        <Link className="nav__menu__link" href="/welcome">
+          Bienvenida
+        </Link>
+      </li>
+      <li>
+        <Link className="nav__menu__link" href="/">
+          Mi calendario
+        </Link>
+      </li>
+      {/* <li>
+            <Link className="nav__menu__link" href="">Categorías</Link>
+          </li>
+          <li>
+            <Link className="nav__menu__link" href="">Estadísticas</Link>
+          </li>
+          <li>
+            <Link className="nav__menu__link" href="">Mi perfil</Link>
+          </li> */}
+    </ul>
+  );
+}
+
+export default Nav;
